Rename validatedPrompts file to promptUtils
