+++ src/commands/label:add.ts
@@ -5,7 +5,7 @@ import { ParseAndHandleError } from '../errors'
 import { checkCurrentRepo } from '../helpers/checkCurrentRepo'
 import { getGithub } from '../helpers/getGithub'
 import { createLabels } from '../helpers/labels'
-import { validatedPrompt } from '../helpers/validatedPrompts'
+import { validatedPrompt } from '../helpers/promptUtils'
 import { AnsSelectReposForLabel, AnsSelectYesNo } from '../types/Answers'
 import { CommandOptions } from '../types/Config'
 import { LabelKeys } from '../types/Labels'
